split factories and moved voteBoundGetter to project_types
